fixed screenshot timings
